# swagger.yaml
openapi: 3.0.3
info:
  title: Pipu Pro
  description: >-
    # Pipu Pro  api documentation
  version: 1.0.0
  contact: { email: ammuftau74@gmail.com }
servers:
  - url: http://localhost:9000/api/v1
    description: This is my local Server
  - url: https://pipu-pro.onrender.com/api/v1
    description: This is the original api URL
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /company:
    post:
      tags:
        - 🏫 Company
      summary: Add new Company
      description: >-
        The API endpoint allows new company.
      operationId: addNewCompany
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: john
                phone:
                  type: string
                  example: '+234568908798'
                email:
                  type: string
                  example: abc@gmail.com
                password:
                  type: string
                  example: company123
      responses:
        '201':
          description: New Company created
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '500'
            Date:
              schema:
                type: string
                example: Fri, 30 Jun 2023 06:44:09 GMT
            ETag:
              schema:
                type: string
                example: W/"1f4-1q38Sbu25p0UVmeYQmrEv2VrgPg"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            RateLimit-Limit:
              schema:
                type: string
                example: '500'
            RateLimit-Remaining:
              schema:
                type: string
                example: '495'
            RateLimit-Reset:
              schema:
                type: string
                example: '772'
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                  message:
                    type: string
                    example: >-
                      Company added successfully.
    patch:
      tags:
        - 🏫 Company
      summary: update Company
      description: >-
        The API endpoint allows company update their profile.
      operationId: updateCompany
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Veek Design
                email:
                  type: string
                  example: 'olivia@gmail.com'
                country:
                  type: string
                  example: 'Nigeria'
                timezone:
                  type: string
                  example: 'Pacific standard time'
      responses:
        '201':
          description: Company updated
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '500'
            Date:
              schema:
                type: string
                example: Fri, 30 Jun 2023 06:44:09 GMT
            ETag:
              schema:
                type: string
                example: W/"1f4-1q38Sbu25p0UVmeYQmrEv2VrgPg"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            RateLimit-Limit:
              schema:
                type: string
                example: '500'
            RateLimit-Remaining:
              schema:
                type: string
                example: '495'
            RateLimit-Reset:
              schema:
                type: string
                example: '772'
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                  message:
                    type: string
                    example: >-
                      Company updated successfully.
...
  /company/financial-policy:
    get:
      tags:
        - 🏫 Company
      summary: Get company financial policy
      description: >-
        The API endpoint allows you to retrieve the company financial policy
      operationId: getCompanyfinancial-policy
      responses:
        '200':
          description: Get company financial policy
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '443'
            Date:
              schema:
                type: string
                example: Fri, 30 Jun 2023 06:44:47 GMT
            ETag:
              schema:
                type: string
                example: W/"1bb-YVaffq7t8o3of0yDl9g26tprr/8"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            RateLimit-Limit:
              schema:
                type: string
                example: '500'
            RateLimit-Remaining:
              schema:
                type: string
                example: '493'
            RateLimit-Reset:
              schema:
                type: string
                example: '631'
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object    
    post:
      tags:
        - 🏫 Company
      summary: Update financial policy
      description: >-
        The API endpoint allows company update financial policy.
      operationId: updateFinancialPolicy
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                companyId:
                  type: string
                  example: '681e2429c5421ead94648571'
                name:
                  type: string
                  example: john
                description:
                  type: string
                  example: '+234568908798'
                permissions:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                        example: manageProfile
                      isAllowed:
                        type: boolean
                        example: true
      responses:
        '201':
          description: New Company created
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '500'
            Date:
              schema:
                type: string
                example: Fri, 30 Jun 2023 06:44:09 GMT
            ETag:
              schema:
                type: string
                example: W/"1f4-1q38Sbu25p0UVmeYQmrEv2VrgPg"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            RateLimit-Limit:
              schema:
                type: string
                example: '500'
            RateLimit-Remaining:
              schema:
                type: string
                example: '495'
            RateLimit-Reset:
              schema:
                type: string
                example: '772'
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                  message:
                    type: string
                    example: >-
                      Company added successfully.
...